Module solana_entry::entry

The entry module is a fundamental building block of Proof of History. It contains a unique ID that is the hash of the Entry before it, plus the hash of the transactions within it. Entries cannot be reordered, and its field num_hashes represents an approximate amount of time since the last Entry was created.

Structs

Api

Entry

Each Entry contains three pieces of data. The num_hashes field is the number of hashes performed since the previous entry. The hash field is the result of hashing hash from the previous entry num_hashes times. The transactions field points to Transactions that took place shortly before hash was generated.

EntrySigVerificationState

EntrySummary

EntryVerificationState

GpuSigVerificationData

GpuVerificationData

PAR_THREAD_POOL 🔒

VerifyRecyclers

__FieldVisitor 🔒

__Visitor 🔒

Enums

DeviceSigVerificationData

DeviceVerificationData

EntryType

Typed entry to distinguish between transaction and tick entries

EntryVerificationStatus

VerifyAction 🔒

Last action required to verify an entry

__Field 🔒

Constants

FIELDS 🔒

Statics

API 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

Traits

EntrySlice

Functions

__stability 🔒

__static_ref_initialize 🔒

api

compare_hashes 🔒

create_random_ticks

create_ticks

hash_transactions

init 🔒

init_poh

next_entry

Creates the next Tick or Transaction Entry num_hashes after start_hash.

next_entry_mut

next_hash

Creates the hash num_hashes after start_hash. If the transaction contains a signature, the final hash will be a hash of both the previous ID and the signature. If num_hashes is zero and there’s no transaction data, start_hash is returned.

next_versioned_entry

Creates the next Tick or Transaction Entry num_hashes after start_hash.

start_verify_transactions

start_verify_transactions_cpu 🔒

start_verify_transactions_gpu 🔒

verify_transactions

Type Definitions

EntryReceiver

EntrySender