Trait solana_entry::entry::EntrySlice

pub trait EntrySlice {
    // Required methods
    fn verify_cpu(&self, start_hash: &Hash) -> EntryVerificationState;
    fn verify_cpu_generic(&self, start_hash: &Hash) -> EntryVerificationState;
    fn verify_cpu_x86_simd(
        &self,
        start_hash: &Hash,
        simd_len: usize
    ) -> EntryVerificationState;
    fn start_verify(
        &self,
        start_hash: &Hash,
        recyclers: VerifyRecyclers
    ) -> EntryVerificationState;
    fn verify(&self, start_hash: &Hash) -> bool;
    fn verify_tick_hash_count(
        &self,
        tick_hash_count: &mut u64,
        hashes_per_tick: u64
    ) -> bool;
    fn tick_count(&self) -> u64;
}

Required Methods

fn verify_cpu(&self, start_hash: &Hash) -> EntryVerificationState

Verifies the hashes and counts of a slice of transactions are all consistent.

fn verify_cpu_generic(&self, start_hash: &Hash) -> EntryVerificationState

fn verify_cpu_x86_simd( &self, start_hash: &Hash, simd_len: usize ) -> EntryVerificationState

fn start_verify( &self, start_hash: &Hash, recyclers: VerifyRecyclers ) -> EntryVerificationState

fn verify(&self, start_hash: &Hash) -> bool

fn verify_tick_hash_count( &self, tick_hash_count: &mut u64, hashes_per_tick: u64 ) -> bool

Checks that each entry tick has the correct number of hashes. Entry slices do not necessarily end in a tick, so tick_hash_count is used to carry over the hash count for the next entry slice.

fn tick_count(&self) -> u64

Counts tick entries